test(core): move guard-registry tests into registries.test.ts
Code review feedback: registry tests are co-located in registries.test.ts
by convention. Also extracts the act expression in the two error-path
tests so it's only evaluated once per assertion.

Also exports WorkflowGuard and InMemoryGuardRegistry from the public
index so the test can import from src/index.js per project convention.
